requirements.txt
streamlit>=1.28.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.2
python-dotenv>=1.0.0

